文件系统层与文件

+++ source/kernel/fs/fs.c
@@ -19,10 +19,12 @@
 
 static list_t mounted_list;             // 挂载链表
 static list_t free_list;                // 空闲链表
-static fs_t fs_table[FS_TABLE_SIZE];    // fs表
+static fs_t fs_table[FS_TABLE_SIZE];    // 文件系统表, 最多能挂载10个文件系统
+                                        // 初始化阶段将其组织为空闲链表
 static fs_t* root_fs;                   // 根文件系统
 
-extern fs_op_t devfs_op;                // TODO: 临时的
+// 文件系统回调结构
+extern fs_op_t devfs_op;                // DEV
 extern fs_op_t fat16fs_op;              // FAT16
 
 void fs_init() {
@@ -30,7 +32,7 @@ void fs_init() {
     // 初始化文件系统的链表
     mount_list_init();
 
-    // 初始化文件表
+    // 初始化系统文件表
     file_table_init();
 
     // 初始化磁盘
@@ -45,35 +47,32 @@ void fs_init() {
     ASSERT(root_fs != NULL);
 }
 
-// TODO: TEMP
-static int is_path_valid(const char* path) {
-    if((path == NULL) || (path[0] == '\0')) {
-        return 0;
-    }
-    return 1;
-}
-
 int sys_unlink(const char* filename) {
-    fs_lock(root_fs);
-    int ret = root_fs->op->unlink(root_fs, filename);
-    fs_unlock(root_fs);
+    int ret = -1;
+    if(root_fs->op->unlink != NULL) {
+        fs_lock(root_fs);
+        ret = root_fs->op->unlink(root_fs, filename);
+        fs_unlock(root_fs);
+    }
     return ret;
 }
 
+// TODO: 若重复打开同一个文件, 会不会多在sys_file_alloc多分配一块?
 int sys_open(const char* filename, int flags, ...) {
-    // 分配文件描述符链接
+    // 再系统文件表中分配一个文件
     file_t* file = file_alloc();
     if(file == NULL) {
         goto sys_open_failed;
     }
 
-    // 分配文件描述符
+    // 在任务中分配文件描述符
     int fd = task_alloc_fd(file);
     if(fd < 0) {
         goto sys_open_failed;
     }
 
-    // 检查名称是否以挂载点开头, 若没有, 则认为filename在根目录下
+    // 检查名称是否以挂载点开头(解析file对应的fs)
+    // 若fs为空, 则认为filename在根目录下
     fs_t* fs = NULL;
     list_node_t* node = list_first(&mounted_list);
     while(node) {
@@ -96,19 +95,21 @@ int sys_open(const char* filename, int flags, ...) {
     file->fs = fs;
     kernel_strncpy(file->filename, filename, FILENAME_SIZE);
 
-    fs_lock(fs);
-    int err = fs->op->open(fs, filename, file);
-    if(err < 0) {
+    if(fs->op->open != NULL) {
+        fs_lock(fs);
+        int ret = fs->op->open(fs, filename, file);
+        if(ret < 0) {
+            fs_unlock(fs);
+            log_print("open %s failed.", filename);
+            goto sys_open_failed;
+        }
         fs_unlock(fs);
-        log_print("open %s failed.", filename);
-        goto sys_open_failed;
     }
-    fs_unlock(fs);
     return fd;
 
 sys_open_failed:
     if(file) {
-        file_free(file);
+        file_dec_ref(file);
     }
     if(fd >= 0) {
         task_remove_fd(fd);
@@ -133,11 +134,13 @@ int sys_read(int fd, char* buf, int len) {
         return -1;
     }
 
+    int ret = -1;
     fs_t* fs = p->fs;
-    
-    fs_lock(fs);
-    int ret = fs->op->read(buf, len, p);
-    fs_unlock(fs);
+    if(fs->op->read != NULL) {
+        fs_lock(fs);
+        ret = fs->op->read(buf, len, p);
+        fs_unlock(fs);
+    }
     return ret;
 }
 
@@ -158,16 +161,17 @@ int sys_write(int fd, char* buf, int len) {
         return -1;
     }
 
+    int ret = -1;
     fs_t* fs = p->fs;
-    
-    fs_lock(fs);
-    int ret = fs->op->write(buf, len, p);
-    fs_unlock(fs);
+    if(fs->op->write != NULL) {
+        fs_lock(fs);
+        ret = fs->op->write(buf, len, p);
+        fs_unlock(fs);
+    }
     return ret;
 }
 
 int sys_lseek(int fd, int offset, int dir) {
-    // TODO:
     if(is_fd_bad(fd)) {
         log_print("file error");
         return -1;
@@ -179,16 +183,17 @@ int sys_lseek(int fd, int offset, int dir) {
         return -1;
     }
 
+    int ret = -1;
     fs_t* fs = p->fs;
-    fs_lock(fs);
-    int ret = fs->op->seek(p, offset, dir);
-    fs_unlock(fs);
+    if(fs->op->seek != NULL) {
+        fs_lock(fs);
+        ret = fs->op->seek(p, offset, dir);
+        fs_unlock(fs);
+    }
     return ret;
 }
 
 int sys_close(int fd) {
-    // TODO: 
-
     if(is_fd_bad(fd)) {
         log_print("file error");
         return -1;
@@ -200,18 +205,20 @@ int sys_close(int fd) {
         return -1;        
     }
 
-    // 要打开文件, 那么断言文件是打开的
+    // 断言文件已经打开
     ASSERT(p->ref > 0);
 
     // 减少ref
-    if(--p->ref == 0) {
+    if(--p->ref == 0) { // 若为0, 那么关闭文件
         fs_t* fs = p->fs;
-        fs_lock(fs);
-        fs->op->close(p);
-        fs_unlock(fs);
-
-        file_free(p);
+        if(fs->op->close != NULL) {
+            fs_lock(fs);
+            fs->op->close(p);
+            fs_unlock(fs);
+        }
+        file_dec_ref(p);
     }
+    // 释放fd
     task_remove_fd(fd);
     return 0;
 }
@@ -245,9 +252,12 @@ int sys_fstat(int fd, struct stat* st) {
     fs_t* fs = p->fs;
     kernel_memset(st, 0, sizeof(struct stat));
 
-    fs_lock(fs);
-    int ret = fs->op->stat(p, st);
-    fs_unlock(fs);
+    int ret = -1;
+    if(fs->op->stat != NULL) {
+        fs_lock(fs);
+        ret = fs->op->stat(p, st);
+        fs_unlock(fs);
+    }
     return ret;
 }
 
@@ -258,18 +268,18 @@ int sys_ioctl(int fd, int cmd, int arg0, int arg1) {
     }
 
     file_t* p = get_task_file(fd);
-    if(!p) {
+    if(p == NULL) {
         log_print("file not opened.");
         return -1;
     }
 
-    fs_t* fs = p->fs;
     int ret = -1;
-    fs_lock(fs);
+    fs_t* fs = p->fs;
     if(fs->op->ioctl != NULL) {
+        fs_lock(fs);
         ret = fs->op->ioctl(p, cmd, arg0, arg1);
+        fs_unlock(fs);
     }
-    fs_unlock(fs);
     return ret;
 }
 
@@ -280,53 +290,63 @@ int sys_dup(int fd) {
     }
 
     file_t* p = get_task_file(fd);
-    if(!p) {
+    if(p == NULL) {
         log_print("file not opened.");
         return -1;
     }
 
-    // 新创建一个fd与p相同
+    // 新增new指向fd对应的文件
     int new = task_alloc_fd(p);
     if(new < 0) {
         log_print("no task file avaliable.");
         return -1;
     }
     
+    // 增加文件的引用计数
     file_inc_ref(p);
     return new;
 }
 
 int sys_opendir(const char* path, DIR* dir) {
-    fs_lock(root_fs);
-    int ret = root_fs->op->opendir(root_fs, path, dir);
-    fs_unlock(root_fs);
+    int ret = -1;
+    if(root_fs->op->opendir != NULL) {
+        fs_lock(root_fs);
+        ret = root_fs->op->opendir(root_fs, path, dir);
+        fs_unlock(root_fs);
+    }
     return ret;
 }
 
 int sys_readdir(DIR* dir, struct dirent* dirent) {
-    fs_lock(root_fs);
-    int ret = root_fs->op->readdir(root_fs, dir, dirent);
-    fs_unlock(root_fs);
+    int ret = -1;
+    if(root_fs->op->readdir != NULL) {
+        fs_lock(root_fs);
+        ret = root_fs->op->readdir(root_fs, dir, dirent);
+        fs_unlock(root_fs);
+    }
     return ret;
 }
 
 int sys_closedir(DIR* dir) {
-    fs_lock(root_fs);
-    int ret = root_fs->op->closedir(root_fs, dir);
-    fs_unlock(root_fs);
+    int ret = -1;
+    if(root_fs->op->closedir != NULL) {
+        fs_lock(root_fs);
+        ret = root_fs->op->closedir(root_fs, dir);
+        fs_unlock(root_fs);
+    }
     return ret;
 }
 
+
 static void mount_list_init() {
-    
-    // 初始化空闲链表
+    // 初始化空闲与挂载链表
+    list_init(&mounted_list);
     list_init(&free_list);
+
+    // 将系统文件表组织成链表
     for(int i = 0; i < FS_TABLE_SIZE; i++) {
         list_insert_first(&free_list, &fs_table[i].node);
     }
-
-    // 初始化挂载链表
-    list_init(&mounted_list);
 }
 
 static fs_op_t* get_fs_op(fs_type_t type, int major) {
@@ -343,12 +363,22 @@ static fs_op_t* get_fs_op(fs_type_t type, int major) {
 }
 
 static fs_t* mount(fs_type_t type, char* mount_point, int dev_major, int dev_minor) {
+
+    /**
+     * 主要流程:
+     *      1. 判断当前挂载点没有被挂载过
+     *      2. 分配一个空闲的文件系统表
+     *      3. 初始化文件系统
+     *      4. 执行挂载
+     */
+
     fs_t* fs = NULL;
     log_print("mount file system, name: %s, dev: %x,", mount_point, dev_major);
 
-    // 判断当前文件是否挂载过
-    list_node_t* curr = list_first(&mounted_list);      // 遍历挂载链表
+    // 1. 遍历挂载链表, 判断当前文件是否挂载过
+    list_node_t* curr = list_first(&mounted_list);
     while(curr != NULL) {
+        // 若fs的挂载点与入参的挂载点一样就退出
         fs_t* p = list_entry_of(curr, fs_t, node);
         if(kernel_strncmp(fs->mount_point, mount_point, FS_MOUNTP_SIZE) == 0) {
             log_print("fs already mounted.");
@@ -357,7 +387,7 @@ static fs_t* mount(fs_type_t type, char* mount_point, int dev_major, int dev_min
         curr = list_node_next(curr);
     }
 
-    // 从空闲链表中分配新的fs
+    // 2. 从空闲链表中分配新的fs
     list_node_t* free_node = list_remove_first(&free_list);
     if(free_node == NULL) {
         log_print("no free fs, mount failed.");
@@ -365,27 +395,25 @@ static fs_t* mount(fs_type_t type, char* mount_point, int dev_major, int dev_min
     }
     fs = list_entry_of(free_node, fs_t, node);
 
-    // 获取当前fs的操作函数
-    fs_op_t* op = get_fs_op(type, dev_major);
-    if(op == NULL) {
-        log_print("unsupport fs type: %d", type);
-        goto mount_failed;
-    }
-
-    // 初始化fs
+    // 3. 初始化fs
     kernel_memset(fs, 0, sizeof(fs_t));
     kernel_strncpy(fs->mount_point, mount_point, FS_MOUNTP_SIZE);
+    
+    fs_op_t* op = get_fs_op(type, dev_major);   // 获取当前fs的操作函数
+    if(op == NULL || op->mount == NULL) {
+        log_print("unsupport fs type: %d or no mount function.", type);
+        goto mount_failed;
+    }
     fs->op = op;
     fs->mtx = NULL;
 
-    // 执行挂载
+    // 4. 执行挂载, 调用fs各自的挂载函数
     if(op->mount(fs, dev_major, dev_minor) < 0) {
         log_print("mount fs %s failed.", mount_point);
         goto mount_failed;
     }
+    list_insert_last(&mounted_list, &fs->node);     // 将节点移至挂载链表中
 
-    // 将节点移至挂载链表中
-    list_insert_last(&mounted_list, &fs->node);
     return fs;
 
 mount_failed:
@@ -404,11 +432,13 @@ const char* path_next_child(const char* path) {
     return *p ? p : NULL;
 }
 
-int path_to_num(const char* path, int* num) {
+int string_to_int(const char* path, int* num) {
     int x = 0;
-
     const char* p = path;
     while(*p) {
+        if(*p < '0' || *p > '9') {
+            return -1;
+        }
         x = x * 10 + (*p - '0');
         p++;
     }
